hud: modal title padding in RTY
